+++ src/main/java/com/cloudogu/scm/review/pullrequest/service/MergeService.java
@@ -26,6 +26,8 @@ package com.cloudogu.scm.review.pullrequest.service;
 import com.cloudogu.scm.review.BranchProtectionHook;
 import com.cloudogu.scm.review.PermissionCheck;
 import com.cloudogu.scm.review.pullrequest.dto.MergeCommitDto;
+import org.apache.shiro.SecurityUtils;
+import sonia.scm.repository.Changeset;
 import sonia.scm.repository.InternalRepositoryException;
 import sonia.scm.repository.NamespaceAndName;
 import sonia.scm.repository.Repository;
@@ -39,14 +41,20 @@ import sonia.scm.repository.api.MergeStrategyNotSupportedException;
 import sonia.scm.repository.api.RepositoryService;
 import sonia.scm.repository.api.RepositoryServiceFactory;
 import sonia.scm.repository.spi.MergeConflictResult;
+import sonia.scm.user.User;
+import sonia.scm.user.UserDisplayManager;
 
 import javax.inject.Inject;
 import java.io.IOException;
 import java.text.MessageFormat;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.Scanner;
 import java.util.Set;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import static com.cloudogu.scm.review.pullrequest.service.PullRequestStatus.OPEN;
@@ -69,13 +77,15 @@ public class MergeService {
   private final PullRequestService pullRequestService;
   private final Collection<MergeGuard> mergeGuards;
   private final BranchProtectionHook branchProtectionHook;
+  private final UserDisplayManager userDisplayManager;
 
   @Inject
-  public MergeService(RepositoryServiceFactory serviceFactory, PullRequestService pullRequestService, Set<MergeGuard> mergeGuards, BranchProtectionHook branchProtectionHook) {
+  public MergeService(RepositoryServiceFactory serviceFactory, PullRequestService pullRequestService, Set<MergeGuard> mergeGuards, BranchProtectionHook branchProtectionHook, UserDisplayManager userDisplayManager) {
     this.serviceFactory = serviceFactory;
     this.pullRequestService = pullRequestService;
     this.mergeGuards = mergeGuards;
     this.branchProtectionHook = branchProtectionHook;
+    this.userDisplayManager = userDisplayManager;
   }
 
   public void merge(NamespaceAndName namespaceAndName, String pullRequestId, MergeCommitDto mergeCommitDto, MergeStrategy strategy, boolean emergency) {
@@ -89,7 +99,7 @@ public class MergeService {
         () -> {
           MergeCommandBuilder mergeCommand = repositoryService.getMergeCommand();
           isAllowedToMerge(repositoryService.getRepository(), mergeCommand, strategy, emergency);
-          prepareMergeCommand(mergeCommand, pullRequest, mergeCommitDto, strategy);
+          prepareMergeCommand(repositoryService, mergeCommand, pullRequest, mergeCommitDto, strategy);
           MergeCommandResult mergeCommandResult = mergeCommand.executeMerge();
 
           if (!mergeCommandResult.isSuccess()) {
@@ -184,11 +194,7 @@ public class MergeService {
       if (RepositoryPermissions.read(repositoryService.getRepository()).isPermitted() && repositoryService.isSupported(Command.LOG)) {
         try {
           StringBuilder builder = new StringBuilder();
-          repositoryService.getLogCommand()
-            .setBranch(pullRequest.getSource())
-            .setAncestorChangeset(pullRequest.getTarget())
-            .getChangesets()
-            .getChangesets()
+          getChangesetsFromLogCommand(pullRequest, repositoryService)
             .forEach(c -> {
               builder.append("- ").append(c.getDescription()).append('\n');
               if (c.getDescription().contains("\n")) {
@@ -212,6 +218,14 @@ public class MergeService {
     }
   }
 
+  private List<Changeset> getChangesetsFromLogCommand(PullRequest pullRequest, RepositoryService repositoryService) throws IOException {
+    return repositoryService.getLogCommand()
+      .setBranch(pullRequest.getSource())
+      .setAncestorChangeset(pullRequest.getTarget())
+      .getChangesets()
+      .getChangesets();
+  }
+
   private void assertPullRequestIsOpen(Repository repository, PullRequest pullRequest) {
     if (pullRequest.getStatus() != OPEN) {
       throw new CannotMergeNotOpenPullRequestException(repository, pullRequest);
@@ -229,11 +243,103 @@ public class MergeService {
     }
   }
 
-  private void prepareMergeCommand(MergeCommandBuilder mergeCommand, PullRequest pullRequest, MergeCommitDto mergeCommitDto, MergeStrategy strategy) {
+  private void prepareMergeCommand(RepositoryService repositoryService, MergeCommandBuilder mergeCommand, PullRequest pullRequest, MergeCommitDto mergeCommitDto, MergeStrategy strategy) {
     mergeCommand.setBranchToMerge(pullRequest.getSource());
     mergeCommand.setTargetBranch(pullRequest.getTarget());
-    mergeCommand.setMessageTemplate(mergeCommitDto.getCommitMessage());
+    String enrichedCommitMessage = enrichCommitMessageWithTrailers(repositoryService, pullRequest, mergeCommitDto, strategy);
+    mergeCommand.setMessageTemplate(enrichedCommitMessage);
     mergeCommand.setMergeStrategy(strategy);
+  }
+
+  private String enrichCommitMessageWithTrailers(RepositoryService repositoryService, PullRequest pullRequest, MergeCommitDto mergeCommitDto, MergeStrategy strategy) {
+    StringBuilder builder = new StringBuilder();
+    builder.append(mergeCommitDto.getCommitMessage());
+
+    List<String> approvers = getPullRequestApprovers(pullRequest);
+    if (shouldAppendTrailers(strategy, approvers)) {
+      builder.append("\n\n");
+
+      String merger = SecurityUtils.getSubject().getPrincipals().oneByType(User.class).getDisplayName();
+      appendSquashCoAuthorsToCommitMessage(repositoryService, pullRequest, strategy, merger, builder);
+      appendApproversToCommitMessage(builder, approvers, merger);
+    }
+    return builder.toString();
+  }
+
+  private void appendSquashCoAuthorsToCommitMessage(RepositoryService repositoryService, PullRequest pullRequest, MergeStrategy strategy, String merger, StringBuilder builder) {
+
+    if (strategy == MergeStrategy.SQUASH) {
+      List<Changeset> changesets = getAllChangesetsForBranch(repositoryService, pullRequest);
+      Set<String> appendedCoAuthorMails = new HashSet<>();
+      appendCommitAuthorsAsCoAuthors(changesets, merger, appendedCoAuthorMails, builder);
+      extractCoAuthorsFromSquashedCommitMessages(changesets, merger, appendedCoAuthorMails, builder);
+    }
+  }
+
+  private void appendCommitAuthorsAsCoAuthors(List<Changeset> changesets, String merger, Set<String> appendedCoAuthorMails, StringBuilder builder) {
+    changesets
+      .stream()
+      .map(Changeset::getAuthor)
+      .distinct()
+      .filter(author -> !author.getName().equals(merger))
+      .forEach(author -> {
+        appendedCoAuthorMails.add(author.getMail());
+        builder.append("Co-authored-by: ").append(author.getName()).append(" <").append(author.getMail()).append(">\n");
+      });
+  }
+
+  private void extractCoAuthorsFromSquashedCommitMessages(List<Changeset> changesets, String author, Set<String> appendedCoAuthors, StringBuilder builder) {
+    Pattern delimiter = Pattern.compile("[\\n]");
+    for (Changeset changeset : changesets) {
+      if (changeset.getDescription() != null) {
+        try (Scanner scanner = new Scanner(changeset.getDescription())) {
+          scanner.useDelimiter(delimiter);
+          while (scanner.hasNext()) {
+            String line = scanner.next();
+            if (shouldExtractLineFromSquashedCommit(appendedCoAuthors, line, author)) {
+              appendedCoAuthors.add(line);
+              builder.append(line).append("\n");
+            }
+          }
+        }
+      }
+    }
+  }
+
+  private boolean shouldExtractLineFromSquashedCommit(Set<String> extractedSquashedCoAuthors, String line, String author) {
+    return line.contains("Co-authored-by") && !line.contains(author) && extractedSquashedCoAuthors.stream().noneMatch(line::contains);
+  }
+
+  private List<Changeset> getAllChangesetsForBranch(RepositoryService repositoryService, PullRequest pullRequest) {
+    try {
+      return getChangesetsFromLogCommand(pullRequest, repositoryService);
+    } catch (IOException e) {
+      throw new InternalRepositoryException(repositoryService.getRepository(), "Could not read changesets from branch " + pullRequest.getSource(), e);
+    }
+  }
+
+  private boolean shouldAppendTrailers(MergeStrategy strategy, List<String> approvers) {
+    return !approvers.isEmpty() || strategy == MergeStrategy.SQUASH;
+  }
+
+  private void appendApproversToCommitMessage(StringBuilder builder, List<String> approvers, String merger) {
+    for (String approver : approvers) {
+      userDisplayManager.get(approver)
+        .ifPresent(reviewer -> {
+          if (!reviewer.getDisplayName().equals(merger)) {
+            builder.append("Reviewed-by: ").append(reviewer.getDisplayName()).append(" <").append(reviewer.getMail()).append(">\n");
+          }
+        });
+    }
+  }
+
+  private List<String> getPullRequestApprovers(PullRequest pullRequest) {
+    return pullRequest.getReviewer()
+      .entrySet()
+      .stream()
+      .filter(Map.Entry::getValue)
+      .map(Map.Entry::getKey)
+      .collect(Collectors.toList());
   }
 
   private MergeCommandBuilder prepareDryRun(RepositoryService repositoryService, String sourceBranch, String targetBranch) {